Move new endpoint to dataset API
