--- superset/datasource/schemas.py
@@ -1,28 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-from marshmallow import fields, Schema
-
-
-class GetOrCreateTableSchema(Schema):
-    table_name = fields.String(required=True, description="Name of table")
-    database_id = fields.Integer(
-        required=True, description="ID of database table belongs to"
-    )
-    schema = fields.String(
-        description="The schema the table belongs to", allow_none=True
-    )
-    template_params = fields.String(description="Template params for the table")
